After Merging the pull request from @imran-prog

+++ _Core_Components_Cpp/functions/maths_func.cpp
@@ -26,9 +26,9 @@ double tan_number(double main_num)
     return result;
 }
 
-int round_number(double main_num)
+double round_number(double main_num)
 {
-    int result = round(main_num);
+    double result = round(main_num);
     return result;
 }
